clean up and relocate leaderboard functionality
