+++ ironforgedbot/commands/leaderboard/leaderboard_registry.py
@@ -1,0 +1,8 @@
+from ironforgedbot.commands.leaderboard.leaderboard_ingots import INGOTS_LEADERBOARD
+from ironforgedbot.commands.leaderboard.leaderboard_score import SCORE_LEADERBOARD
+from ironforgedbot.commands.leaderboard.leaderboard_types import LeaderboardConfig
+
+LEADERBOARD_TYPES: dict[str, LeaderboardConfig] = {
+    "ingots": INGOTS_LEADERBOARD,
+    "score": SCORE_LEADERBOARD,
+}
